% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A549_TAD}
\alias{A549_TAD}
\title{A549_TAD file}
\format{
A data frame with 1866 rows and 3 variables:
\describe{
\item{\code{chromosome}}{character chromosome location of TADs}
\item{\code{start}}{integer 1-indexed start position of TADs}
\item{\code{end}}{integer 1-indexed end position of TADs}
}
}
\source{
\url{http://3dgenome.fsm.northwestern.edu/publications.html}
}
\usage{
A549_TAD
}
\description{
A bed-3 file with a collection of hg38-annotated topologically associated domains from A549 cell lines, created by the Yue lab (3D genome browser)
}
\keyword{datasets}
