% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AEC_donor_day0_H3K27ac}
\alias{AEC_donor_day0_H3K27ac}
\title{Day 0 donor AEC H3K27ac ChIP-seq peak dataset}
\format{
A data frame with 126069 rows and 3 variables:
\describe{
\item{\code{chromosome}}{character chromosome location of peaks}
\item{\code{start}}{integer 1-indexed start position of peaks}
\item{\code{end}}{integer 1-indexed start position of peaks}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM4250941}
}
\usage{
AEC_donor_day0_H3K27ac
}
\description{
A bed-3 file with hg38-annotated H3K27ac ChIP-seq peaks from AEC cells derived day 0 from donor lung (i.e. type2-like)
}
\keyword{datasets}
