% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bedlike_overlap_function.R
\name{bedlike_overlap_function}
\alias{bedlike_overlap_function}
\title{bedlike_overlap_function}
\usage{
bedlike_overlap_function(
  dataset_to_annotate,
  dataset_to_overlap_with,
  within_distance = 0,
  overlap_type = "partial"
)
}
\arguments{
\item{dataset_to_annotate}{A data frame or GRanges object, consisting of regions for which the user wants to find overlaps in the dataset_to_overlap_with dataset. This dataset should have bed-like formatting, with the first column containing the chromosome each region is located on, the second column containing the 1-indexed start position of each region, and the third column containing the 1-indexed end position of each region.}

\item{dataset_to_overlap_with}{A second data frame or GRanges object, which contains the regions the user wishes to record that overlap with each of the regions in the dataset_to_annotate. This dataset should also have bed-like formatting, with the first column containing the chromosome each region is located on, the second column containing the 1-indexed start position of each region, and the third column containing the 1-indexed end position of each region.}

\item{within_distance}{Specify a buffer size which is effectively added to the regions in the dataset_to_overlap_with to calculate potential overlaps i.e., positive values specify the buffer region in which overlaps between the two can be identified, even if they would not be found normally. Defaults to 0.}

\item{overlap_type}{Specifies the type of overlaps the user would like to identify. Valid options include "partial" which still counts overlaps if only part of the regions in the dataset_to_overlap_with dataset overlap with each region in the dataset_to_annotate dataset, and "complete", which requires that each region in the dataset_to_overlap_with dataset completely enclose a region in the dataset_to_annotate dataset for an overlap to be counted between the two. Defaults to "partial".}
}
\value{
Returns a vector with the length of the number of regions in the dataset_to_annotate dataset, with the rownames of the dataset_to_overlap_with that each region of the dataset_to_annotate dataset was found to overlap with.
}
\description{
This function takes two bed-like data frames or GRanges objects as primary arguments,
the first three columns of which contain chromosome, start, and end locations, respectively.
The function then returns the rownames of the regions in the "dataset_to_overlap_with" which
overlap with each of the regions in the "dataset_to_annotate".
}
\examples{
# Generate two example datasets:
annotate <- data.frame(
   "chromosome"= c("chr1","chr1","chr1"),
   "start"= c(50,150,400),
   "end"= c(100,200,500),
   stringsAsFactors= FALSE
)

overlap <- data.frame(
   "chromosome"= c("chr1","chr1"),
   "start"= c(25,130,190),
   "end"= c(125,160,250),
   stringsAsFactors= FALSE
)

# Perform overlapping with defaults
bedlike_overlap_function(
   dataset_to_annotate= annotate,
   dataset_to_overlap_with= overlap
)

}
